Renamed main src directory to ps_typer and made it a python package (also updated imports)
