+++ ps_typer/assets/InconsolataBold.ttf
