+++ ps_typer/source_ui/result_window.py
