+++ ps_typer/source_ui/settingsWindow.ui
