+++ ps_typer/source_ui/typing_window.py
