+++ ps_typer/type_test/texts.py
